ll_covers/requirements.txt
gspread>=5.12.0
google-auth>=2.23.0
yt-dlp>=2024.0.0
